requirements.txt
Flask
boto3
Flask-JWT-Extended
pytest
